+++ views/room.ejs
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>videoChatApp</title>
-    <link rel="stylesheet" href="room.css" />
+    <title>Video Chat</title>
+    <link rel="stylesheet" href="/Video Chat/video.css" />
     <script src="/socket.io/socket.io.js"></script>
     <script src="https://kit.fontawesome.com/c939d0e917.js"></script>
     <script src="https://unpkg.com/peerjs@1.3.1/dist/peerjs.min.js"></script>
@@ -62,5 +62,5 @@
     </div>
   </div>
   </body>
-  <script src="room.js"></script>
+  <script src="/Video Chat/video.js"></script>
 </html>
